update mobile user
 - login, logout, register user
 - mapp api post

+++ Mobile/src/components/Register.js
@@ -3,20 +3,32 @@ import { View, Text, TextInput, TouchableOpacity, StyleSheet } from "react-nativ
 import AsyncStorage from '@react-native-async-storage/async-storage';
 import axios from 'axios';
 
-const Login = ({ navigation }) => {
-  const [email, setEmail] = useState('');
-  const [password, setPassword] = useState('');
+const Register = ({ navigation }) => {
+  const [username, setUsername] = useState("");
+  const [email, setEmail] = useState("");
+  const [password, setPassword] = useState("");
+  const [confirmpassword, setConfirmPassword] = useState("");
   const [error, setError] = useState("");
-
   const loginHandler = async (e) => {
     e.preventDefault();
     setError('')
-
+    if (password !== confirmpassword) {
+      setPassword("");
+      setConfirmPassword("");
+      setTimeout(() => {
+        setError("");
+      }, 8000);
+      return setError("Passwords do not match");
+    }
     try {
       const URI = 'http://localhost:5000'
       const { data } = await axios.post(
-        `${URI}/auth/login`,
-        { email, password }
+        `${URI}/auth/register`,
+        {
+          username,
+          email,
+          password,
+        }
       );
       await AsyncStorage.setItem("authToken", data.token);
 
@@ -25,7 +37,6 @@ const Login = ({ navigation }) => {
         navigation.navigate("Profile")
 
       }, 1800)
-
 
     } catch (error) {
       setError(error.response.data.error);
@@ -45,21 +56,39 @@ const Login = ({ navigation }) => {
       </View>
       <View style={styles.inputView} >
         <TextInput
-          value={password}
-          secureTextEntry={true}
           style={styles.inputText}
+          placeholder="Username"
+          placeholderTextColor="#003f5c"
+          value={username}
+          onChangeText={setUsername}
+        />
+      </View>
+      <View style={styles.inputView} >
+        <TextInput
+          value={password}
+          style={styles.inputText}
+          secureTextEntry={true}
           placeholder="Mật khẩu"
           placeholderTextColor="#003f5c"
           onChangeText={setPassword} />
       </View>
+      <View style={styles.inputView} >
+        <TextInput
+          value={confirmpassword}
+          style={styles.inputText}
+          secureTextEntry={true}
+          placeholder="Xác nhập mật khẩu"
+          placeholderTextColor="#003f5c"
+          onChangeText={setConfirmPassword} />
+      </View>
       <Text style={styles.textError}>
         {error}
       </Text>
-      <TouchableOpacity onPress={() => navigation.navigate('Register')}>
-        <Text style={styles.forgot}>Tạo tài khoản</Text>
+      <TouchableOpacity>
+        <Text style={styles.forgot} onPress={() => navigation.navigate('Profile')}>Đã có tài khoản</Text>
       </TouchableOpacity>
       <TouchableOpacity style={styles.loginBtn} onPress={loginHandler}>
-        <Text style={styles.loginText}>Đăng nhập</Text>
+        <Text style={styles.loginText}>Đăng ký</Text>
       </TouchableOpacity>
     </View>
   );
@@ -91,11 +120,13 @@ const styles = StyleSheet.create({
   },
   inputText: {
     height: 50,
+    fontWeight: "600",
     color: "#0063a5"
   },
   forgot: {
     color: "#0063a5",
-    fontSize: 11
+    fontSize: 11,
+    fontWeight: "600",
   },
   loginBtn: {
     width: "80%",
@@ -108,6 +139,7 @@ const styles = StyleSheet.create({
     marginBottom: 10
   },
   loginText: {
+    fontWeight: "600",
     color: "white"
   },
   textError: {
@@ -115,5 +147,5 @@ const styles = StyleSheet.create({
   }
 });
 
-export default Login;
+export default Register;
